+++ permissions_auditor/__init__.py
@@ -1,6 +1,6 @@
 "Permissions auditing for Django."
 
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 
 # This can be removed once Django 3.1 and below support is dropped.
 default_app_config = 'permissions_auditor.apps.PermissionsAuditorConfig'